moved Sendgrid to core logic as per design diagram
