+++ src/test/java/teammates/test/cases/logic/EmailsTest.java
@@ -38,12 +38,12 @@ import teammates.logic.core.CoursesLogic;
 import teammates.logic.core.Emails;
 import teammates.logic.core.FeedbackSessionsLogic;
 import teammates.logic.core.InstructorsLogic;
+import teammates.logic.core.Sendgrid;
 import teammates.logic.core.StudentsLogic;
 import teammates.test.cases.BaseComponentTestCase;
 import teammates.test.cases.ui.browsertests.SystemErrorEmailReportTest;
 import teammates.test.driver.AssertHelper;
 import teammates.test.driver.TestProperties;
-import teammates.googleSendgridJava.Sendgrid;
 
 public class EmailsTest extends BaseComponentTestCase {
